bmi to double in chw

+++ lib/screens/chw/new_patient_questionnairs/new_patient_questionnaire_screen.dart
@@ -1236,8 +1236,8 @@ class _MeasurementsState extends State<Measurements> {
 
   calculateBmi() {
     if (heightEditingController.text != '' && weightEditingController.text != '') {
-      var height = int.parse(heightEditingController.text) / 100;
-      var weight = int.parse(weightEditingController.text);
+      var height = double.parse(heightEditingController.text) / 100;
+      var weight = double.parse(weightEditingController.text);
 
       var bmi = weight / (height * height);
